Revert "LevelProvider :lol]"
This reverts commit c5e40fac828f726ab11b777707bb4c5aecea4b6e.

+++ src/main/java/cn/nukkit/nbt/CompoundTag.java
@@ -109,7 +109,7 @@ public class CompoundTag extends Tag {
     }
 
     public int getInt(String name) {
-        if (!tags.containsKey(name)) return 0;
+        if (!tags.containsKey(name)) return (int) 0;
         return ((IntTag) tags.get(name)).data;
     }
